refactor: Add title parameter to CustomTextMazadDetails and update HomeDetails screens
This commit refactors the CustomTextMazadDetails widget to accept a `title` parameter, allowing for dynamic title text. It also updates the HomeDetails screens (Jaraa, Sayantaliq, Qadim, and Muntahi) to pass specific titles to the CustomTextMazadDetails widget for "تفاصيل المزاد" (Auction Details) and "تفاصيل المنتج" (Product Details) sections. This change improves the reusability and flexibility of the CustomTextMazadDetails widget and provides more context-specific information on the Home Details screens.