feat: Rename Sayantaliq to Ready for clarity in ongoing auctions
This commit renames all instances of "Sayantaliq" to "Ready" to provide better clarity and consistency across the application, specifically for auctions that are in the ready state. This change affects the data models, repositories, cubits, UI components, API services, and routing configurations related to these auctions.

The following changes were made:

- Renamed `Sayantaliq` to `Ready` in data models (`sayantaliq_auction_response.dart`, `sayantaliq_show_auction_mode.dart`).
- Updated repository names from `SayantaliqAutionRepo` and `SayantaliqShowAutionRepo` to `ReadyAutionRepo` and `ReadyShowAutionRepo` respectively.
- Renamed cubits and states from `SayantaliqCubit` and `SayantaliqState` to `ReadyCubit` and `ReadyState` respectively. Also, `SayantaliqShowAustionCubit` and `SayantaliqShowAustionState` to `ReadyShowAustionCubit` and `ReadyShowAustionState` respectively.
- Updated UI components to reflect the new naming convention (`custom_sayantilq_card_item.dart`, `home_details_sayantaliq_screen.dart`, `custom_sayantaliq_list_view.dart`).
- Updated API service calls to use the new `Ready` terminology.
- Updated route names to use `Ready` instead of `Sayantaliq` (`app_routes.dart`, `route.dart`).
- Updated references in `home_screen.dart` and `handleNotification.dart`.

These changes improve the overall understandability and maintainability of the codebase by using more descriptive and consistent naming for auctions that are in the ready state.